requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==5.0.0
flask-restx==1.3.0
Werkzeug==3.0.6
python-dotenv==1.0.0
insightface>=0.7.3
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
qrcode[pil]>=7.4.2
pyzbar>=0.1.9